requirements.txt
streamlit>=1.28.0
google-genai>=1.0.0
Pillow>=10.0.0
tenacity>=8.2.0
